refactor: unir las entidades pack y producto en una unica entidad
