--- src/main/java/com/application/persistence/repository/PackRepository.java
@@ -1,13 +1,0 @@
-package com.application.persistence.repository;
-
-import com.application.persistence.entity.pack.Pack;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-
-@Repository
-public interface PackRepository extends JpaRepository<Pack, Long> {
-    List<Pack> findByActivoTrue();
-    List<Pack> findByCategorias_CategoriaId(Long categoriaId);
-}
