+++ src/main/java/com/application/persistence/repository/ProductoRepository.java
@@ -9,4 +9,5 @@ import java.util.List;
 @Repository
 public interface ProductoRepository extends JpaRepository<Producto, Long> {
     List<Producto> findByActivoTrue();
+    List<Producto> findByCategorias_CategoriaId(Long categoriaId);
 }
