+++ src/main/java/com/application/presentation/dto/pack/response/PackResponse.java
@@ -1,6 +1,6 @@
 package com.application.presentation.dto.pack.response;
 
-import com.application.persistence.entity.pack.enums.ETipo;
+import com.application.persistence.entity.producto.enums.ETipo;
 
 import java.util.List;
